【update】参加認証画面作成
